Refactoring of k-means clustering code
Moved to separate file and transparently handle k-means++ initialization in
wrapper function instead of in caller.
